[ten] Support local mode
