+++ packages/ten/src/components/LocalGameCard.tsx
@@ -1,0 +1,7 @@
+import React, { ReactElement } from 'react';
+
+import GameCardWithLogic from './GameCardWithLogic';
+
+export default function PlayAgainstAIGameCard(): ReactElement {
+  return <GameCardWithLogic />;
+}
