--- packages/ten/src/components/LocalGameCards.module.css
@@ -1,8 +1,0 @@
-.LocalGameCardsContainer {
-  display: flex;
-  justify-content: center;
-}
-
-.LocalGameCardsContainer > div {
-  margin: 1em;
-}
