+++ packages/ten/src/pages/local.tsx
@@ -1,14 +1,14 @@
 import React, { ReactElement } from 'react';
 
 import App from '../components/App';
-import LocalGameCards from '../components/LocalGameCards';
+import GameCardWithLogic from '../components/GameCardWithLogic';
 
 import './index.css';
 
 export default function Local(): ReactElement {
   return (
     <App>
-      <LocalGameCards />
+      <GameCardWithLogic />
     </App>
   );
 }
